requirement.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pandasai
langchain_google_genai
openai
unstructured